requirements.txt
asn1crypto==0.23.0
astroid==1.5.3
backports.functools-lru-cache==1.4
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.0.1
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
chatbase==0.2
click==6.7
configparser==3.5.0
cryptography==2.0.3
dill==0.2.7.1
Django==1.11.4
django-cors-headers==2.1.0
django-orientdb-api==0.1
djangorestframework==3.6.4
djangorestframework-csv==2.0.0
enum34==1.1.6
Flask==0.12.2
future==0.16.0
futures==3.1.1
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gapic-google-cloud-pubsub-v1==0.15.4
gapic-google-cloud-spanner-admin-database-v1==0.15.3
gapic-google-cloud-spanner-admin-instance-v1==0.15.3
gapic-google-cloud-spanner-v1==0.15.3
google-api-core==0.1.1
google-api-python-client==1.6.2
google-auth==1.2.0
google-cloud==0.27.0
google-cloud-bigquery==0.26.0
google-cloud-bigtable==0.26.0
google-cloud-core==0.26.0
google-cloud-datastore==1.2.0
google-cloud-dns==0.26.0
google-cloud-error-reporting==0.26.0
google-cloud-language==0.30.0
google-cloud-logging==1.2.0
google-cloud-monitoring==0.26.0
google-cloud-pubsub==0.27.0
google-cloud-resource-manager==0.26.0
google-cloud-runtimeconfig==0.26.0
google-cloud-spanner==0.26.0
google-cloud-speech==0.28.0
google-cloud-storage==1.3.2
google-cloud-translate==1.1.0
google-cloud-videointelligence==0.25.0
google-cloud-vision==0.26.0
google-gax==0.15.15
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
googlemaps==2.5.1
grpc-google-iam-v1==0.11.4
grpcio==1.7.0
gspread==0.6.2
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
lxml==4.0.0
MarkupSafe==1.0
mccabe==0.6.1
monotonic==1.4
oauth2client==3.0.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-pubsub-v1==0.15.4
proto-google-cloud-spanner-admin-database-v1==0.15.3
proto-google-cloud-spanner-admin-instance-v1==0.15.3
proto-google-cloud-spanner-v1==0.15.3
protobuf==3.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
pyee==3.0.3
pylint==1.7.2
pyOpenSSL==17.3.0
pyorient==1.5.5
pytz==2017.2
regex==2017.7.28
requests==2.18.4
rsa==3.4.2
singledispatch==3.4.0.3
six==1.11.0
slackclient==1.0.9
slackeventsapi==1.0.0
tenacity==4.5.0
unicodecsv==0.14.1
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
websocket-client==0.44.0
Werkzeug==0.12.2
wrapt==1.10.11
